decimal to uint256

+++ {{cookiecutter.project_name}}/contracts/NFT.vy
@@ -148,7 +148,7 @@ MAX_SUPPLY: constant(uint256) = {{cookiecutter.max_supply_amount}}
 
 {%- if cookiecutter.royalties == 'y' %}
 # @dev Percentage of royalties for lifetime for the creator
-ROYALTY_PERCENTAGE: constant(decimal) = {{cookiecutter.royalty_percentage}}
+ROYALTY_PERCENTAGE: constant(uint256) = {{cookiecutter.royalty_percentage}}
 ROYALTY_TO_APPLY_TO_PRICE: constant(decimal) = {{ ROYALTY_PERCENTAGE / 100.0 }}
 {%- endif %}
